Check for file existence before removing

+++ src/Krevindiou/BagheeraBundle/Entity/Account.php
@@ -218,7 +218,7 @@ class Account
      */
     public function removeUpload()
     {
-        if ($file = $this->getAbsolutePath()) {
+        if (($file = $this->getAbsolutePath()) && is_file($file)) {
             unlink($file);
         }
     }
